+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## unreleased
 
+## v0.2.1 (2026-06-13)
+
 - feat: removed reddit and table-el styles
 - feat: renamed -p/--decimal-places to -N/--decimal-places
 - feat: migrated heavy-outline, orgtbl, table-el, rst-grid styles to tabled